+++ lib/features/home/presentation/views/web/home_view.dart
@@ -1,15 +1,13 @@
 // ignore: prefer_const_constructors
 import 'package:flutter/material.dart';
-import 'package:flutter_svg/svg.dart';
 import 'package:get/get.dart';
-import 'package:google_fonts/google_fonts.dart';
 import 'package:thepos/features/carts/presentation/controllers/carts_controller.dart';
-import 'package:thepos/features/carts/presentation/views/cart_view.dart';
+import 'package:thepos/features/carts/presentation/views/web/cart_view.dart';
 import 'package:thepos/features/home/data/models/product.dart';
 import 'package:thepos/features/home/presentation/controllers/home_controller.dart';
-import 'package:thepos/features/home/presentation/widgets/category_widget.dart';
-import 'package:thepos/features/home/presentation/widgets/header_home.dart';
-import 'package:thepos/features/home/presentation/widgets/product_widget.dart';
+import 'package:thepos/features/home/presentation/widgets/web/category_widget.dart';
+import 'package:thepos/features/home/presentation/widgets/web/header_home.dart';
+import 'package:thepos/features/home/presentation/widgets/web/product_widget.dart';
 
 class HomeView extends GetView<HomeController> {
   HomeView({Key? key}) : super(key: key);
