+++ lib/features/home/presentation/widgets/mobile/categories_widget.dart
